avant pull eliott

+++ include/Application.h
@@ -20,7 +20,7 @@ public:
     ~Application() {}
 
     // Méthodes
-    std::map<std::string, float> moyenneQualiteAir(float latitude, float longitude, time_t debut, time_t fin, float perimetre) const;
+    std::map<std::string, int> moyenneQualiteAir(float latitude, float longitude, time_t debut, time_t fin, float perimetre) const;
     float estimerQualiteAir(float latitude, float longitude) const;
     std::vector<std::pair<Capteur, float>> listerCapteursSimilaires(Capteur &capteur) const;
     void ajouterPointUtilisateur(Utilisateur &user) const;
